+++ src/core/mfa.ts
@@ -23,15 +23,43 @@
  *
  */
 
-import { createTOTP } from '../../src/core/totp';
+import { TOTP } from "otpauth";
 
-describe('TOTP', () => {
-  it('generate one time token', () => {
-    const totp = createTOTP({ secret: "FLIIOLP3IR3W" });
-    expect(totp.generate().length).toBe(6);
-    expect(totp.toString()).toEqual('otpauth://totp/SyntheticsTOTP?secret=FLIIOLP3IR3Q&algorithm=SHA1&digits=6&period=30')
-  });
-});
+type TOTPOptions = {
+  /**
+   * Provider or Service the secret is associated with
+   */
+  issuer?: string
+  /**
+   * Account Identifier.
+   * @default 'SyntheticsTOTP'
+   */
+  label?: string
+  /**
+   * Include issuer prefix in label.
+   */
+  issuerInLabel?: boolean
+  /**
+   * The encoded secret key used to generate the TOTP.
+   */
+  secret?: string
+  /**
+   * The algorithm used to generate the TOTP.
+   * @default 'SHA1'
+   */
+  algorithm?: string
+  /**
+   * Number of digits in the generated token.
+   * @default 6
+   */
+  digits?: number
+  /**
+   * Validity period in seconds for the token.
+   * @default 30
+   */
+  period?: number
+};
 
-
-
+export function totp(secret?: string, options: TOTPOptions = {}) {
+  return new TOTP({ label: "SyntheticsTOTP", secret, ...options }).generate();
+}
